# docs/swagger.yaml
definitions:
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        format: int64
        type: integer
    type: object
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
...
  user.User:
    properties:
      gid:
        description: |-
          Gid is the primary group ID.
          On POSIX systems, this is a decimal number representing the gid.
          On Windows, this is a SID in a string format.
          On Plan 9, this is the contents of /dev/user.
        type: string
      homeDir:
        description: HomeDir is the path to the user's home directory (if they have
          one).
        type: string
      name:
        description: |-
          Name is the user's real or display name.
          It might be blank.
          On POSIX systems, this is the first (or only) entry in the GECOS field
          list.
          On Windows, this is the user's display name.
          On Plan 9, this is the contents of /dev/user.
        type: string
      uid:
        description: |-
          Uid is the user ID.
          On POSIX systems, this is a decimal number representing the uid.
          On Windows, this is a security identifier (SID) in a string format.
          On Plan 9, this is the contents of /dev/user.
        type: string
      username:
        description: Username is the login name.
        type: string
    type: object
...
  wn_internal_domain_dto.ChangePictureResponse:
    properties:
      newImgUrl:
        type: string
    type: object
  wn_internal_domain_dto.ChangeProfilePicture:
    properties:
      file:
        $ref: '#/definitions/multipart.FileHeader'
      userId:
        type: string
    required:
    - file
    type: object
...
  wn_pkg_response.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/wn_pkg_response.meta'
      pagination:
        $ref: '#/definitions/wn_pkg_response.pagination'
    type: object
  wn_pkg_response.meta:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      requestId:
        type: string
    type: object
  wn_pkg_response.pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      perPage:
        type: integer
    type: object
info:
  contact: {}
  description: This is walrus notes api service.
  title: WALRUS NOTES API
  version: "1.0"
paths:
...
  /wn/api/v1/user/picture:
    post:
      description: сменить аватарку пользователя
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.ChangeProfilePicture'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.ChangePictureResponse'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: user_not_found'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: change_profile_picture
      tags:
      - user
  /wn/api/v1/user/profile/{id}:
    get:
      description: получить юзера по айди
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.User'
              type: object
        "400":
          description: 'possible codes: bind_path, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: user_not_found'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: get_user_by_id
      tags:
      - user
swagger: "2.0"